+++ CMakeLists.txt
@@ -2,7 +2,7 @@ cmake_minimum_required(VERSION 3.10)
 project(ruminate)
 set (CMAKE_CXX_STANDARD 20)
 
-file(GLOB SRC "src/*.cpp" "src/modules/*.cpp")
+file(GLOB_RECURSE SRC "src/*.cpp")
 add_executable(${PROJECT_NAME} ${SRC} main.cpp)
 
 target_include_directories(${PROJECT_NAME} PUBLIC include)
