+++ include/layers.h
@@ -31,5 +31,5 @@ namespace rum
     };
 
     template <typename T>
-    concept LayerType = std::is_base_of<Layer<typename T::type>, T>::value;
-}; // namespace rum
+    concept LayerType = std::is_base_of_v<Layer<typename T::type>, T>;
+} // namespace rum
